Segunda subida

+++ misperrispy/core/models.py
@@ -43,7 +43,7 @@ class Perro(models.Model):
     nombre = models.CharField(max_length=50)
     raza = models.CharField(max_length=100)
     edad = models.IntegerField()
-    tamano = models.DecimalField(max_digits=10, decimal_places=2)
+    tamano = models.IntegerField()
 
     def __str__(self):
         return self.nombre
